refactor(coral): Move team api mocks.
