+++ coral/src/app/features/topics/hooks/teams/useGetTeams.test.tsx
@@ -2,8 +2,8 @@ import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { renderHook, waitFor } from "@testing-library/react";
 import { ReactElement } from "react";
 import { server } from "src/services/api-mocks/server";
-import { mockGetTeams } from "src/domain/topic/topic-api.msw";
 import { useGetTeams } from "src/app/features/topics/hooks/teams/useGetTeams";
+import { mockGetTeams } from "src/domain/team/team-api.msw";
 
 const queryClient = new QueryClient({
   defaultOptions: {
